Moved CustomRouter to extensions.rest_framework_ex
