+++ FAIRshake/settings.py
@@ -75,6 +75,7 @@ INSTALLED_APPS = [
     'extensions.drf_yasg_ex',
     'extensions.rest_auth_ex',
     'extensions.versions_ex',
+    'extensions.rest_framework_ex',
     'FAIRshakeHub',
     'FAIRshakeAPI',
 ]
